chore: refactored tests to reflect internal changes
