+++ bindings/nodejs/examples/1-create-account.js
@@ -4,11 +4,12 @@
 
 require('dotenv').config();
 const { CoinType } = require('../out/types');
-const manager = require('./account-manager');
+const { AccountManager } = require('@iota/wallet');
 
 async function run() {
     try {
-        await manager.setStrongholdPassword(process.env.SH_PASSWORD);
+        const manager = await createAccountManager();
+
         // The coin type only needs to be set on the first account
         const account = await manager.createAccount({
             alias: 'Alice',
@@ -24,6 +25,33 @@ async function run() {
         console.log('Error: ' + error);
     }
     process.exit(0);
+}
+
+async function createAccountManager() {
+    const accountManagerOptions = {
+        storagePath: './alice-database',
+        clientOptions: {
+            nodes: [
+                {
+                    url: 'https://api.alphanet.iotaledger.net',
+                    auth: null,
+                    disabled: false,
+                },
+            ],
+            localPow: true,
+        },
+        secretManager: {
+            Stronghold: {
+                snapshotPath: `./wallet.stronghold`,
+                password: `${process.env.SH_PASSWORD}`,
+            },
+        },
+    };
+
+    const manager = new AccountManager(accountManagerOptions);
+    await manager.storeMnemonic(process.env.MNEMONIC);
+    await manager.setStrongholdPassword(process.env.SH_PASSWORD);
+    return manager;
 }
 
 run();
